Change thresh_fact in LinearGradientBadPixelDetection from an int to a float. The original implementation of this at https://github.jpl.nasa.gov/IMD/BadPixelRepairDetection/tree/master  of this had this as an int. The python version just accidentally support float values because python uses duck typing, although this wasn't intended. However we now have a dependency on the unintended support of float thresh_fact. Easy enough to change the code to support a thresh_fact of float - so we've now changed the C++ code for this.

+++ lib/GeoCalCore/Implementation/misc_array_function.cc
@@ -205,7 +205,7 @@ blitz::Array<bool, 2> GeoCal::linear_gradient_bad_pixel_detection
 (const blitz::Array<double, 2>& Img,
  int Window_size,
  double Percentile,
- int Thresh_fact,
+ double Thresh_fact,
  double Nfail_thresh_percentage,
  array_local_edge_handle Edge_handle,
  int Number_task
